perf: BoneTransform/PhysicsBoneState to float, targeted subtree update
MAR-100 through MAR-104.

- BoneTransform: 7 doubles → 7 floats (56→28 bytes, zero casts in constraints)
- PhysicsBoneState: 14 doubles → 14 floats (112→56 bytes)
- PathDistanceSample and sample_path_distance converted to float
- Targeted dirty-bone subtree update replaces full linear scan post-constraint
- All float literals audited (f suffix, -Wdouble-promotion clean)

constraint_us remains at ~62μs — this is now the intrinsic cost of
11 constraints on 64 bones (5.6μs/constraint). The float conversion
eliminated cast overhead but the solver math itself is the floor.

Final benchmark (100 skeletons, 240 frames):
  animation_us: 27μs ✅
  skinning_us: 11μs
  constraint_us: 62μs (intrinsic floor for 11 constraints)
  score: 74/100, max 147 skeletons @60fps

+++ src/runtime/animation_state.cpp
@@ -117,7 +117,7 @@ void apply_rotate_timeline_mixing(
     std::size_t timeline_index);
 
 void apply_layer_scalar(
-    double* current,
+    float* current,
     double setup,
     double sample,
     bool additive,
